+++ src/server/devices/baseDeviceInterfaces/iHumiditySensor.ts
@@ -3,7 +3,10 @@ import { iRoomDevice } from './iRoomDevice';
 export const UNDEFINED_HUMIDITY_VALUE = -1;
 
 export interface iHumiditySensor extends iRoomDevice {
-  humidity: number;
+  readonly persistHumiditySensorInterval: NodeJS.Timeout;
+  readonly humidity: number;
 
   addHumidityCallback(pCallback: (pValue: number) => void): void;
+
+  persistHumiditySensor(): void;
 }